requirements.txt
aiohttp>=3.8.1, <3.9
pandas>=1.4.2, <2.0
typing-extensions>=4.2.0, <5.0
backports.zoneinfo>=0.2.1, <1.0
pytest>=7.1.2, <8.0
pytest-asyncio>=0.18.0,<1.0
pre-commit>= 2.19.0, < 3.0
freezegun>=1.2.1, <1.3.0
oic>=1.4.0, <1.5.0
python-jose>=3.3, <4.0
cachetools>=5.2.0, <5.3.0
marshmallow-jsonapi>=0.24.0, <1.0
